[build] Add support for building with Dune.
We add preliminary support for building with Dune to lablgtk2; this
seems to work well and should provide some advantages:

- compositional build with other developments, for example CoqIDE,
- it allows for an easier layout / library reorganization if wished,
- automatic management of install targets, opam install file, META, stubs.

I used a more explicit dune-file style for now, that while a bit
verbose, should be very easy to understand.

This is a preliminary PR, but it works to build CoqIDE compositionaly.

One strong motivation to do this PR was to help the GTK 3 porting
effort and this provides a seamless developer experience.

TODO:

- add missing libraries, reorganize?
- make all examples work again,
- cleanup makefiles, remove old make system?
- test on OSX, win32.
- add `(optional ...)` fields.

MINOR:

- generate verbose rules automatically.