[build] Add support for building with Dune.
We add preliminary support for building with Dune to lablgtk3; it
seems to work well and does provide some advantages:

- compositional build with other developments, for example CoqIDE,
- it allows for an easier layout / library reorganization if wished,
- automatic management of install targets, opam install file, META, stubs,
- automatic generation of documentation with `odoc`.

Release / usage instructions are placed in `README.dune.md`.

The dune files use an "explicit" style. While a bit more
verbose than other setup, it is very easy to understand.

Note we pass `-Wno-deprecated-declarations` to the C compiler as to
avoid a super-verbose build. We also disable a few OCaml warnings.

One strong motivation to do this PR is to help the GTK 3 porting
effort and this provides a seamless developer experience.

Following discussion on caml-list, we provide one opam package for
each C library. This is IMO more consistent with what packaging
systems do and allows for better dependencies in the OPAM world. For
example, Coq can depend now on `lablgtk3-sourceview3`.

We still depend on the `conf-packages` as to get depext information,
but it is not necessary as the dune-based config script also does C
version checks.

Things that would be worth to do in future PRs are:

- linting: the large amount of OCaml and C warnings should be addressed
- OSX and Win32 support, Travis CI for these platforms
- make all examples work again
- cleanup makefiles, remove old make system. To be done after we add
  OSX/Win32 testing
- generate `varcc` / `propcc` rules using promote